Release/v2023.01.0

+++ services/auth-server/app/app/views.py
@@ -17,15 +17,23 @@ from flask import (
     send_from_directory,
 )
 from werkzeug.security import check_password_hash, generate_password_hash
+from werkzeug.wrappers import Response as ResponseBase
 
 from app.connections import db
 from app.models import Token, User
 from app.utils import PathType, _AuthCacheDictionary, get_auth_cache, set_auth_cache
+from config import CONFIG_CLASS
 
 # This auth_cache is shared between requests
 # within the same Flask process
 _auth_cache: _AuthCacheDictionary = {}
 _auth_cache_age: int = 3  # in seconds
+
+
+def _logout(response: ResponseBase) -> None:
+    response.delete_cookie("auth_token", samesite="Lax")
+    response.delete_cookie("auth_username", samesite="Lax")
+    response.delete_cookie("auth_user_uuid", samesite="Lax")
 
 
 def register_views(app: Flask) -> None:
@@ -91,9 +99,18 @@ def register_views(app: Flask) -> None:
 
         # Automatically redirect to root if request is authenticated
         if is_authenticated(request) and path == "":
-            return handle_login(redirect_type="server")
+            res = handle_login(redirect_type="server")
+        else:
+            res = serve_static_or_dev(path)
 
-        return serve_static_or_dev(path)
+        # See comment in /auth.
+        if (
+            isinstance(res, ResponseBase)
+            and not app.config["AUTH_ENABLED"]
+            and request.cookies.get("auth_user_uuid")
+        ):
+            _logout(res)
+        return res
 
     @app.route("/login/admin", methods=["GET"])
     def login_admin() -> Tuple[str, int] | Response:
@@ -107,6 +124,13 @@ def register_views(app: Flask) -> None:
     def index() -> Tuple[Literal[""], Literal[200]] | Tuple[Literal[""], Literal[401]]:
         # validate authentication through token
         if is_authenticated(request):
+            # Make sure no auth cookies are there if auth mode is not
+            # enabled. This covers an edge case when setting auth mode
+            # from True to False and allows the FE to correctly take
+            # some decision w.r.t. the presence of cookies. The redirect
+            # to login will cause cookies to be cleared in this case.
+            if not app.config["AUTH_ENABLED"] and request.cookies.get("auth_user_uuid"):
+                return "", 401
             return "", 200
         else:
             return "", 401
@@ -114,8 +138,7 @@ def register_views(app: Flask) -> None:
     @app.route("/login/clear", methods=["GET"])
     def logout() -> Response | None:
         resp = redirect_response("/")
-        resp.set_cookie("auth_token", "", samesite="Lax")
-        resp.set_cookie("auth_username", samesite="Lax")
+        _logout(resp)
         return resp
 
     def redirect_response(url: str, redirect_type: str = "server") -> Response:
@@ -185,6 +208,7 @@ def register_views(app: Flask) -> None:
                     # samesite="Lax" to avoid CSRF attacks.
                     resp.set_cookie("auth_token", token.token, samesite="Lax")
                     resp.set_cookie("auth_username", username, samesite="Lax")
+                    resp.set_cookie("auth_user_uuid", user.uuid, samesite="Lax")
 
                     return resp
 
@@ -212,6 +236,24 @@ def register_views(app: Flask) -> None:
                 elif self_username == to_delete_username:
                     return jsonify({"error": "Deleting own user is not allowed."}), 405
                 else:
+                    resp = requests.delete(
+                        (
+                            f"http://{CONFIG_CLASS.ORCHEST_API_ADDRESS}/api/"
+                            f"auth-users/{user.uuid}"
+                        )
+                    )
+                    if resp.status_code != 200:
+                        return (
+                            jsonify(
+                                {
+                                    "error": (
+                                        "Failed to delete auth-user reference in "
+                                        "orchest-api."
+                                    )
+                                }
+                            ),
+                            500,
+                        )
                     db.session.delete(user)
                     db.session.commit()
                     return ""
@@ -251,6 +293,21 @@ def register_views(app: Flask) -> None:
                 uuid=str(uuid.uuid4()),
             )
 
+            resp = requests.post(
+                f"http://{CONFIG_CLASS.ORCHEST_API_ADDRESS}/api/auth-users/",
+                json={"uuid": user.uuid},
+            )
+            if resp.status_code != 201:
+                return (
+                    jsonify(
+                        {
+                            "error": (
+                                "Failed to create auth-user reference in orchest-api."
+                            )
+                        }
+                    ),
+                    500,
+                )
             db.session.add(user)
             db.session.commit()
             return ""
